Fix somw sign conversion bugs

+++ include/sdb/ht.h
@@ -32,7 +32,7 @@ static inline ut32 sdbkv_value_len(const SdbKv *kv) {
 	return kv->base.value_len;
 }
 
-SDB_API SdbKv* sdbkv_new2(const char *k, int kl, const char *v, int vl);
+SDB_API SdbKv* sdbkv_new2(const char *k, size_t kl, const char *v, size_t vl);
 SDB_API SdbKv* sdbkv_new(const char *k, const char *v);
 extern SDB_API void sdbkv_free(SdbKv *kv);
